Add .todense() method to Dask SparseFrame
It works by mapping SparseFrame.todense onto partitions.
It as necessary to allow `map_partitions` to return other types
then SparseFrame, so kwarg `cls` was added. It implies that one cannot
use `cls` kwarg as an argument to mapped function (because it will be
consumed by `map_partitions` and not passed to a mapped function).